Pat2Feat update
