+++ CHANGELOG.txt
@@ -3,6 +3,13 @@ CHANGELOG
 =========
 
 -------------------------------------------------------------------------------
+May, 20, 2022 1.3.2
+-------------------------------------------------------------------------------
+
+Minor:
+- Refactor encodings generation component as Pat2Feat
+
+-------------------------------------------------------------------------------
 April, 28, 2022 1.3.1
 -------------------------------------------------------------------------------
