+++ docs/_sources/quick.rst.txt
@@ -14,12 +14,12 @@ Constraint-based Sequential Pattern Mining
     # Seq2Pat over 3 sequences
     seq2pat = Seq2Pat(sequences=[["A", "A", "B", "A", "D"],
                                  ["C", "B", "A"],
-                         ["C", "A", "C", "D"]])
+                                 ["C", "A", "C", "D"]])
 
     # Price attribute corresponding to each item
     price = Attribute(values=[[5, 5, 3, 8, 2],
-                      [1, 3, 3],
-                      [4, 5, 2, 1]])
+                              [1, 3, 3],
+                              [4, 5, 2, 1]])
 
     # Average price constraint
     seq2pat.add_constraint(3 <= price.average() <= 4)
@@ -34,8 +34,8 @@ Dichotomic Pattern Mining
     # Example to show how to run Dichotomic Pattern Mining
     # on sequences with positive and negative outcomes
     from sequential.seq2pat import Seq2Pat
-    from sequential.dpm import (get_one_hot_encodings, dichotomic_pattern_mining,
-                                DichotomicAggregation)
+    from sequential.pat2feat import Pat2Feat
+    from sequential.dpm import dichotomic_pattern_mining, DichotomicAggregation
 
     # Create seq2pat model for positive sequences
     sequences_pos = [["A", "A", "B", "A", "D"]]
@@ -55,5 +55,6 @@ Dichotomic Pattern Mining
 
     # Encodings of all sequences
     sequences = sequences_pos + sequences_neg
-    encodings = get_one_hot_encodings(sequences, dpm_patterns,
+    pat2feat = Pat2Feat()
+    encodings = pat2feat.get_features(sequences, dpm_patterns,
                                       drop_pattern_frequency=False)
