+++ docs/index.html
@@ -102,12 +102,12 @@ and the Tepper School of Business at CMU.</p>
 <span class="c1"># Seq2Pat over 3 sequences</span>
 <span class="n">seq2pat</span> <span class="o">=</span> <span class="n">Seq2Pat</span><span class="p">(</span><span class="n">sequences</span><span class="o">=</span><span class="p">[[</span><span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;B&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;D&quot;</span><span class="p">],</span>
                              <span class="p">[</span><span class="s2">&quot;C&quot;</span><span class="p">,</span> <span class="s2">&quot;B&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">],</span>
-                     <span class="p">[</span><span class="s2">&quot;C&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;C&quot;</span><span class="p">,</span> <span class="s2">&quot;D&quot;</span><span class="p">]])</span>
+                             <span class="p">[</span><span class="s2">&quot;C&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;C&quot;</span><span class="p">,</span> <span class="s2">&quot;D&quot;</span><span class="p">]])</span>
 
 <span class="c1"># Price attribute corresponding to each item</span>
 <span class="n">price</span> <span class="o">=</span> <span class="n">Attribute</span><span class="p">(</span><span class="n">values</span><span class="o">=</span><span class="p">[[</span><span class="mi">5</span><span class="p">,</span> <span class="mi">5</span><span class="p">,</span> <span class="mi">3</span><span class="p">,</span> <span class="mi">8</span><span class="p">,</span> <span class="mi">2</span><span class="p">],</span>
-                  <span class="p">[</span><span class="mi">1</span><span class="p">,</span> <span class="mi">3</span><span class="p">,</span> <span class="mi">3</span><span class="p">],</span>
-                  <span class="p">[</span><span class="mi">4</span><span class="p">,</span> <span class="mi">5</span><span class="p">,</span> <span class="mi">2</span><span class="p">,</span> <span class="mi">1</span><span class="p">]])</span>
+                          <span class="p">[</span><span class="mi">1</span><span class="p">,</span> <span class="mi">3</span><span class="p">,</span> <span class="mi">3</span><span class="p">],</span>
+                          <span class="p">[</span><span class="mi">4</span><span class="p">,</span> <span class="mi">5</span><span class="p">,</span> <span class="mi">2</span><span class="p">,</span> <span class="mi">1</span><span class="p">]])</span>
 
 <span class="c1"># Average price constraint</span>
 <span class="n">seq2pat</span><span class="o">.</span><span class="n">add_constraint</span><span class="p">(</span><span class="mi">3</span> <span class="o">&lt;=</span> <span class="n">price</span><span class="o">.</span><span class="n">average</span><span class="p">()</span> <span class="o">&lt;=</span> <span class="mi">4</span><span class="p">)</span>
@@ -120,8 +120,8 @@ and the Tepper School of Business at CMU.</p>
 <div class="highlight-python notranslate"><div class="highlight"><pre><span></span><span class="c1"># Example to show how to run Dichotomic Pattern Mining</span>
 <span class="c1"># on sequences with positive and negative outcomes</span>
 <span class="kn">from</span> <span class="nn">sequential.seq2pat</span> <span class="kn">import</span> <span class="n">Seq2Pat</span>
-<span class="kn">from</span> <span class="nn">sequential.dpm</span> <span class="kn">import</span> <span class="p">(</span><span class="n">get_one_hot_encodings</span><span class="p">,</span> <span class="n">dichotomic_pattern_mining</span><span class="p">,</span>
-                            <span class="n">DichotomicAggregation</span><span class="p">)</span>
+<span class="kn">from</span> <span class="nn">sequential.pat2feat</span> <span class="kn">import</span> <span class="n">Pat2Feat</span>
+<span class="kn">from</span> <span class="nn">sequential.dpm</span> <span class="kn">import</span> <span class="n">dichotomic_pattern_mining</span><span class="p">,</span> <span class="n">DichotomicAggregation</span>
 
 <span class="c1"># Create seq2pat model for positive sequences</span>
 <span class="n">sequences_pos</span> <span class="o">=</span> <span class="p">[[</span><span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;B&quot;</span><span class="p">,</span> <span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="s2">&quot;D&quot;</span><span class="p">]]</span>
@@ -141,7 +141,8 @@ and the Tepper School of Business at CMU.</p>
 
 <span class="c1"># Encodings of all sequences</span>
 <span class="n">sequences</span> <span class="o">=</span> <span class="n">sequences_pos</span> <span class="o">+</span> <span class="n">sequences_neg</span>
-<span class="n">encodings</span> <span class="o">=</span> <span class="n">get_one_hot_encodings</span><span class="p">(</span><span class="n">sequences</span><span class="p">,</span> <span class="n">dpm_patterns</span><span class="p">,</span>
+<span class="n">pat2feat</span> <span class="o">=</span> <span class="n">Pat2Feat</span><span class="p">()</span>
+<span class="n">encodings</span> <span class="o">=</span> <span class="n">pat2feat</span><span class="o">.</span><span class="n">get_features</span><span class="p">(</span><span class="n">sequences</span><span class="p">,</span> <span class="n">dpm_patterns</span><span class="p">,</span>
                                   <span class="n">drop_pattern_frequency</span><span class="o">=</span><span class="kc">False</span><span class="p">)</span>
 </pre></div>
 </div>
