+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["api","examples","index","installation","quick"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,"sphinx.ext.intersphinx":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["api.rst","examples.rst","index.rst","installation.rst","quick.rst"],objects:{"sequential.seq2pat":[[0,1,1,"","Attribute"],[0,1,1,"","Seq2Pat"]],"sequential.seq2pat.Attribute":[[0,2,1,"","average"],[0,2,1,"","gap"],[0,2,1,"","median"],[0,2,1,"","span"],[0,3,1,"","values"]],"sequential.seq2pat.Seq2Pat":[[0,2,1,"","add_constraint"],[0,2,1,"","get_patterns"],[0,2,1,"","remove_constraint"],[0,3,1,"","sequences"]],sequential:[[0,0,0,"-","seq2pat"]]},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","property","Python property"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:property"},terms:{"0":0,"1":[0,2,4],"2":[2,4],"2019":2,"2022":1,"3":[2,3,4],"4":[2,4],"5":[2,4],"6":3,"8":[2,4],"class":0,"float":0,"import":[2,3,4],"int":0,"public":2,"return":0,A:[0,2,4],For:3,If:[0,3],It:2,On:3,The:[0,2,3],To:3,_baseconstraint:0,aaai:2,across:2,add:0,add_constraint:[0,2,4],advantag:2,aggreg:[2,4],aggregation_to_pattern:[2,4],al:2,algorithm:2,all:[0,2,3,4],alreadi:0,also:3,an:[0,1,2],analyz:1,ani:3,anoth:3,api:2,approach:2,ar:0,art:2,artifact:3,attribut:[0,2,3,4],avail:1,averag:[0,2,4],b:[2,4],backend:2,base:[0,1,2,4],bdist_wheel:3,between:[0,1,2],bring:2,build:2,build_ext:3,busi:2,c:[2,3,4],cach:3,can:[0,1,3],cd:3,clang:3,cmu:2,collabor:2,command:3,compat:3,compil:3,compli:3,confirm:3,consecut:[0,2],constraint:[0,1,4],correl:1,correspond:[2,4],creat:[2,3,4],current:3,cython:[2,3],d:[2,4],databas:[2,4],decis:2,decreas:0,defin:0,depend:3,desir:2,develop:2,diagram:2,dichotom:[1,2,4],dichotomic_pattern_min:[2,4],dichotomicaggreg:[2,4],differ:[0,1,2],dir:3,directori:3,discov:[2,3],dist:3,doe:0,don:3,downstream:1,dpm:[1,2,4],dpm_pattern:[2,4],drop_pattern_frequ:[2,4],e:[0,3],each:[0,2,4],effici:2,element:0,embed:1,encod:[2,4],enforc:0,et:2,event:[0,2],event_1:0,event_2:0,event_3:0,event_n:0,everi:[0,2],exampl:[2,3,4],exist:0,express:2,extens:3,fals:[2,4],fidel:2,find:[2,4],first:0,folder:3,follow:3,forget:3,form:0,found:1,frequenc:0,frequent:[0,1,2,4],from:[0,2,4],gap:[0,2],gcc:3,gener:[0,1],get:[2,4],get_one_hot_encod:[2,4],get_pattern:[0,2,4],git:3,github:2,given:[0,2,4],handl:0,high:2,hosseininasab:2,host:2,how:[1,2,3,4],i:[0,3],implement:2,includ:3,index:2,inner:0,inplac:3,instal:2,integ:0,integr:1,interfac:2,invest:2,item:[2,4],joint:2,jupyt:[],kadioglu:1,keyerror:0,larg:2,last:0,latest:3,least:[2,4],length:0,level:2,librari:0,line:3,list:0,low:2,m:3,mac:3,make:3,master:3,maximum:[0,2],median:[0,2],min_frequ:[0,2,4],min_frequency_neg:[2,4],min_frequency_po:[2,4],mine:[0,1,2,4],minimum:[0,2],mismatch:0,model:[1,2,4],modul:2,more:3,most:0,ms:3,multi:2,neg:[2,4],none:3,noreturn:0,notebook:[1,3],now:3,num:0,number:[0,3],object:0,occur:[0,2,4],oper:0,origin:3,outcom:[1,2,4],over:[2,4],packag:3,pass:3,pattern:[0,1,4],percentag:0,perform:0,perspect:2,pip:3,plai:1,platform:3,pleas:3,posit:[2,4],price:[2,4],project:3,properti:[0,2],propos:1,pull:3,py3:3,py:3,pypi:3,python:[2,3],reason:2,refer:3,remov:0,remove_constraint:0,replac:3,repo:3,repres:0,requir:2,research:2,restrict:0,role:1,row:0,run:[1,2,3,4],s:1,same:3,school:2,section:3,see:3,seq2pat:[1,3,4],seq2pat_neg:[2,4],seq2pat_po:[2,4],sequenc:[0,1,4],sequences_neg:[2,4],sequences_po:[2,4],sequenti:[0,1,2,3,4],set:[],set_valu:[],setup:2,setuptool:3,shell:3,should:[0,3],show:[2,4],sore:0,sourc:0,span:[0,2],specifi:2,state:2,store:0,string:0,studio:3,subject:[2,4],success:3,support:[1,2],sure:3,t:3,take:2,task:1,tepper:2,test:2,thi:[0,2,3],togeth:2,tool:3,twice:[2,4],two:[0,2,3],txt:3,typeerror:0,under:3,underli:3,union:[0,2,4],unittest:3,upgrad:2,us:[1,3],usag:[2,3],valu:[0,2,4],valueerror:0,version:3,visual:3,wa:3,wang:1,we:1,wheel:3,where:0,whl:3,window:3,written:2,x:[1,3],you:3,your:2},titles:["Seq2Pat Public API","Usage Examples","Seq2Pat: Sequence-to-Pattern Generation Library","Installation","Quick Start"],titleterms:{"public":0,api:0,avail:2,build:3,code:[2,3],constraint:2,content:2,exampl:1,from:3,gener:2,indic:2,instal:3,librari:[2,3],option:3,pattern:2,quick:[2,4],requir:3,seq2pat:[0,2],sequenc:2,setup:3,sourc:[2,3],start:[2,4],tabl:2,test:3,upgrad:3,usag:1,your:3}})
+Search.setIndex({docnames:["api","examples","index","installation","quick"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,"sphinx.ext.intersphinx":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["api.rst","examples.rst","index.rst","installation.rst","quick.rst"],objects:{"sequential.seq2pat":[[0,1,1,"","Attribute"],[0,1,1,"","Seq2Pat"]],"sequential.seq2pat.Attribute":[[0,2,1,"","average"],[0,2,1,"","gap"],[0,2,1,"","median"],[0,2,1,"","span"],[0,3,1,"","values"]],"sequential.seq2pat.Seq2Pat":[[0,2,1,"","add_constraint"],[0,2,1,"","get_patterns"],[0,2,1,"","remove_constraint"],[0,3,1,"","sequences"]],sequential:[[0,0,0,"-","seq2pat"]]},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","property","Python property"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:property"},terms:{"0":0,"1":[0,2,4],"2":[2,4],"2019":2,"2022":1,"3":[2,3,4],"4":[2,4],"5":[2,4],"6":3,"8":[2,4],"class":0,"float":0,"import":[2,3,4],"int":0,"public":2,"return":0,A:[0,2,4],For:3,If:[0,3],It:2,On:3,The:[0,2,3],To:3,_baseconstraint:0,aaai:2,across:2,add:0,add_constraint:[0,2,4],advantag:2,aggreg:[2,4],aggregation_to_pattern:[2,4],al:2,algorithm:2,all:[0,2,3,4],alreadi:0,also:3,an:[0,1,2],analyz:1,ani:3,anoth:3,api:2,approach:2,ar:0,art:2,artifact:3,attribut:[0,2,3,4],avail:1,averag:[0,2,4],b:[2,4],backend:2,base:[0,1,2,4],bdist_wheel:3,between:[0,1,2],bring:2,build:2,build_ext:3,busi:2,c:[2,3,4],cach:3,can:[0,1,3],cd:3,clang:3,cmu:2,collabor:2,command:3,compat:3,compil:3,compli:3,confirm:3,consecut:[0,2],constraint:[0,1,4],correl:1,correspond:[2,4],creat:[2,3,4],current:3,cython:[2,3],d:[2,4],databas:[2,4],decis:2,decreas:0,defin:0,depend:3,desir:2,develop:2,diagram:2,dichotom:[1,2,4],dichotomic_pattern_min:[2,4],dichotomicaggreg:[2,4],differ:[0,1,2],dir:3,directori:3,discov:[2,3],dist:3,doe:0,don:3,downstream:1,dpm:[1,2,4],dpm_pattern:[2,4],drop_pattern_frequ:[2,4],e:[0,3],each:[0,2,4],effici:2,element:0,embed:1,encod:[2,4],enforc:0,et:2,event:[0,2],event_1:0,event_2:0,event_3:0,event_n:0,everi:[0,2],exampl:[2,3,4],exist:0,express:2,extens:3,fals:[2,4],fidel:2,find:[2,4],first:0,folder:3,follow:3,forget:3,form:0,found:1,frequenc:0,frequent:[0,1,2,4],from:[0,2,4],gap:[0,2],gcc:3,gener:[0,1],get:[2,4],get_featur:[2,4],get_one_hot_encod:[],get_pattern:[0,2,4],git:3,github:2,given:[0,2,4],handl:0,high:2,hosseininasab:2,host:2,how:[1,2,3,4],i:[0,3],implement:2,includ:3,index:2,inner:0,inplac:3,instal:2,integ:0,integr:1,interfac:2,invest:2,item:[2,4],joint:2,jupyt:[],kadioglu:1,keyerror:0,larg:2,last:0,latest:3,least:[2,4],length:0,level:2,librari:0,line:3,list:0,low:2,m:3,mac:3,make:3,master:3,maximum:[0,2],median:[0,2],min_frequ:[0,2,4],min_frequency_neg:[2,4],min_frequency_po:[2,4],mine:[0,1,2,4],minimum:[0,2],mismatch:0,model:[1,2,4],modul:2,more:3,most:0,ms:3,multi:2,neg:[2,4],none:3,noreturn:0,notebook:[1,3],now:3,num:0,number:[0,3],object:0,occur:[0,2,4],oper:0,origin:3,outcom:[1,2,4],over:[2,4],packag:3,pass:3,pat2feat:[2,4],pattern:[0,1,4],percentag:0,perform:0,perspect:2,pip:3,plai:1,platform:3,pleas:3,posit:[2,4],price:[2,4],project:3,properti:[0,2],propos:1,pull:3,py3:3,py:3,pypi:3,python:[2,3],reason:2,refer:3,remov:0,remove_constraint:0,replac:3,repo:3,repres:0,requir:2,research:2,restrict:0,role:1,row:0,run:[1,2,3,4],s:1,same:3,school:2,section:3,see:3,seq2pat:[1,3,4],seq2pat_neg:[2,4],seq2pat_po:[2,4],sequenc:[0,1,4],sequences_neg:[2,4],sequences_po:[2,4],sequenti:[0,1,2,3,4],set:[],set_valu:[],setup:2,setuptool:3,shell:3,should:[0,3],show:[2,4],sore:0,sourc:0,span:[0,2],specifi:2,state:2,store:0,string:0,studio:3,subject:[2,4],success:3,support:[1,2],sure:3,t:3,take:2,task:1,tepper:2,test:2,thi:[0,2,3],togeth:2,tool:3,twice:[2,4],two:[0,2,3],txt:3,typeerror:0,under:3,underli:3,union:[0,2,4],unittest:3,upgrad:2,us:[1,3],usag:[2,3],valu:[0,2,4],valueerror:0,version:3,visual:3,wa:3,wang:1,we:1,wheel:3,where:0,whl:3,window:3,written:2,x:[1,3],you:3,your:2},titles:["Seq2Pat Public API","Usage Examples","Seq2Pat: Sequence-to-Pattern Generation Library","Installation","Quick Start"],titleterms:{"public":0,api:0,avail:2,build:3,code:[2,3],constraint:2,content:2,exampl:1,from:3,gener:2,indic:2,instal:3,librari:[2,3],option:3,pattern:2,quick:[2,4],requir:3,seq2pat:[0,2],sequenc:2,setup:3,sourc:[2,3],start:[2,4],tabl:2,test:3,upgrad:3,usag:1,your:3}})
