+++ docsrc/quick.rst
@@ -34,8 +34,8 @@ Dichotomic Pattern Mining
     # Example to show how to run Dichotomic Pattern Mining
     # on sequences with positive and negative outcomes
     from sequential.seq2pat import Seq2Pat
-    from sequential.dpm import (get_one_hot_encodings, dichotomic_pattern_mining,
-                                DichotomicAggregation)
+    from sequential.pat2feat import Pat2Feat
+    from sequential.dpm import dichotomic_pattern_mining, DichotomicAggregation
 
     # Create seq2pat model for positive sequences
     sequences_pos = [["A", "A", "B", "A", "D"]]
@@ -55,5 +55,6 @@ Dichotomic Pattern Mining
 
     # Encodings of all sequences
     sequences = sequences_pos + sequences_neg
-    encodings = get_one_hot_encodings(sequences, dpm_patterns,
+    pat2feat = Pat2Feat()
+    encodings = pat2feat.get_features(sequences, dpm_patterns,
                                       drop_pattern_frequency=False)
