+++ sequential/_version.py
@@ -2,4 +2,4 @@
 # Copyright FMR LLC <opensource@fidelity.com>
 # SPDX-License-Identifier: GPL-2.0
 
-__version__ = "1.3.1"
+__version__ = "1.3.2"
