+++ tests/test_dpm.py
@@ -5,246 +5,13 @@ import os
 import unittest
 
 from sequential.seq2pat import Seq2Pat, Attribute
-from sequential.dpm import get_one_hot_encodings, dichotomic_pattern_mining, DichotomicAggregation
+from sequential.pat2feat import Pat2Feat
+from sequential.dpm import dichotomic_pattern_mining, DichotomicAggregation
 
 
 class TestDPMUtils(unittest.TestCase):
     TEST_DIR = os.path.dirname(os.path.abspath(__file__))
     DATA_DIR = TEST_DIR + os.sep + "data" + os.sep
-
-    def test_one_hot_encoding_csp_local_with_constraints(self):
-
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        values = [[5, 5, 3, 8, 2],
-                  [1, 3, 3],
-                  [4, 5, 2, 1]]
-
-        # Seq2Pat over 3 sequences
-        seq2pat = Seq2Pat(sequences)
-
-        # Create price attributes
-        price = Attribute(values)
-
-        # Create price constraint
-        price_ct = 3 <= price.median() <= 4
-
-        # Constraint to specify the median of prices in a pattern
-        seq2pat.add_constraint(price_ct)
-
-        # Find sequences with min_frequency=2
-        patterns = seq2pat.get_patterns(min_frequency=2)
-
-        # Create encoding
-        encodings = get_one_hot_encodings(sequences, patterns,
-                                          constraints=[price_ct], drop_pattern_frequency=True)
-        # sequence      feat0
-        # [A,A,B,A,D]    1
-        # [C, B, A]      0
-        # [C, A, C, D]   1
-
-        self.assertListEqual([[1], [0], [1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_global_with_constraints(self):
-
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        values = [[5, 5, 3, 8, 2],
-                  [1, 3, 3],
-                  [4, 5, 2, 1]]
-
-        # Seq2Pat over 3 sequences
-        seq2pat = Seq2Pat(sequences)
-
-        # Create price attributes
-        price = Attribute(values)
-
-        # Create price constraint
-        price_ct = 3 <= price.median() <= 4
-
-        # Constraint to specify the median of prices in a pattern
-        seq2pat.add_constraint(price_ct)
-
-        # Find sequences with min_frequency=2
-        patterns = seq2pat.get_patterns(min_frequency=2)
-
-        # Create encoding with csp_global when rolling_window_size=None
-        encodings = get_one_hot_encodings(sequences, patterns, constraints=[price_ct],
-                                          rolling_window_size=None, drop_pattern_frequency=True)
-        # sequence      feat0
-        # [A,A,B,A,D]    1
-        # [C, B, A]      0
-        # [C, A, C, D]   1
-
-        self.assertListEqual([[1], [0], [1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_local_without_constraints(self):
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        # Seq2Pat over 3 sequences
-        seq2pat = Seq2Pat(sequences)
-
-        # Find sequences with min_frequency=2
-        patterns = seq2pat.get_patterns(min_frequency=2)
-
-        # Create encoding
-        encodings = get_one_hot_encodings(sequences, patterns, drop_pattern_frequency=True)
-        # encoding is a data frame
-        # sequence      feat0 feat1 feat2
-        # [A,A,B,A,D]    1    1     0
-        # [C, B, A]      0    1     1
-        # [C, A, C, D]   1    0     1
-
-        self.assertListEqual([[1, 1, 0], [0, 1, 1], [1, 0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_global_without_constraints(self):
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        # Seq2Pat over 3 sequences
-        seq2pat = Seq2Pat(sequences)
-
-        # Find sequences with min_frequency=2
-        patterns = seq2pat.get_patterns(min_frequency=2)
-
-        # Create encoding
-        encodings = get_one_hot_encodings(sequences, patterns, rolling_window_size=None,
-                                          drop_pattern_frequency=True)
-        # encoding is a data frame
-        # sequence      feat0 feat1 feat2
-        # [A,A,B,A,D]    1    1     0
-        # [C, B, A]      0    1     1
-        # [C, A, C, D]   1    0     1
-
-        self.assertListEqual([[1, 1, 0], [0, 1, 1], [1, 0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_local_random_patterns_without_constraints(self):
-        # Define the sequences and their attributes to create one-hot encoding
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        patterns = [['B', 'D'], ['A', 'C', 'D']]
-
-        # Create encoding, do not need to drop frequency in the end of each pattern
-        encodings = get_one_hot_encodings(sequences, patterns, drop_pattern_frequency=False)
-
-        self.assertListEqual([[1, 0], [0, 0], [0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_global_random_patterns_without_constraints(self):
-        # Define the sequences and their attributes to create one-hot encoding
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        patterns = [['B', 'D'], ['A', 'C', 'D']]
-
-        # Create encoding, do not need to drop frequency in the end of each pattern
-        encodings = get_one_hot_encodings(sequences, patterns, rolling_window_size=None,
-                                          drop_pattern_frequency=False)
-
-        self.assertListEqual([[1, 0], [0, 0], [0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_local_random_patterns_with_constraints(self):
-        # Define the sequences and their attributes to create one-hot encoding
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        attributes = [[5, 5, 3, 8, 2],
-                      [1, 3, 3],
-                      [4, 5, 2, 1]]
-
-        # Create price attributes
-        price = Attribute(values=attributes)
-
-        price_ct = price.median() <= 2
-
-        patterns = [['B', 'D'], ['A', 'C', 'D']]
-
-        # Create encoding
-        encodings = get_one_hot_encodings(sequences, patterns, constraints=[price_ct], drop_pattern_frequency=False)
-
-        self.assertListEqual([[0, 0], [0, 0], [0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_csp_global_random_patterns_with_constraints(self):
-        # Define the sequences and their attributes to create one-hot encoding
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        attributes = [[5, 5, 3, 8, 2],
-                      [1, 3, 3],
-                      [4, 5, 2, 1]]
-
-        # Create price attributes
-        price = Attribute(values=attributes)
-
-        price_ct = price.median() <= 2
-
-        patterns = [['B', 'D'], ['A', 'C', 'D']]
-
-        # Create encoding
-        encodings = get_one_hot_encodings(sequences, patterns, constraints=[price_ct],
-                                          rolling_window_size=None, drop_pattern_frequency=False)
-
-        self.assertListEqual([[0, 0], [0, 0], [0, 1]], encodings.values[:, 1:].tolist())
-
-    def test_one_hot_encoding_string_patterns_without_dropping_frequency(self):
-        # Define the sequences and their attributes to create one-hot encoding
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        attributes = [[5, 5, 3, 8, 2],
-                      [1, 3, 3],
-                      [4, 5, 2, 1]]
-
-        # Create price attributes
-        price = Attribute(values=attributes)
-
-        price_ct = price.median() <= 2
-
-        patterns = [['B', 'D', 1], ['A', 'C', 'D', 1]]
-
-        # Create encoding
-        with self.assertRaises(ValueError):
-            # This should fail when patterns have frequency appended to the end of patterns
-            encodings = get_one_hot_encodings(sequences, patterns, constraints=[price_ct], drop_pattern_frequency=False)
-            print(encodings)
-
-    def test_constraints_as_positional_arg(self):
-        sequences = [["A", "A", "B", "A", "D"],
-                     ["C", "B", "A"],
-                     ["C", "A", "C", "D"]]
-
-        values = [[5, 5, 3, 8, 2],
-                  [1, 3, 3],
-                  [4, 5, 2, 1]]
-
-        # Seq2Pat over 3 sequences
-        seq2pat = Seq2Pat(sequences)
-
-        # Create price attributes
-        price = Attribute(values)
-        # Create price attributes
-        price_ct = 3 <= price.median() <= 4
-        # Constraint to specify the median of prices in a pattern
-        seq2pat.add_constraint(price_ct)
-        # Find sequences with min_frequency=2
-        patterns = seq2pat.get_patterns(min_frequency=2)
-
-        # Create encoding with/without constraints
-        encodings = get_one_hot_encodings(sequences, patterns, [price_ct], drop_pattern_frequency=True)
-
-        self.assertListEqual([[1], [0], [1]], encodings.values[:, 1:].tolist())
 
     def test_dichotomic_pattern_mining_union(self):
         sequences_pos = [["A", "A", "B", "A", "D"],
@@ -486,7 +253,8 @@ class TestDPMUtils(unittest.TestCase):
         price = Attribute(values=values_pos + values_neg)
         price_ct = 3 <= price.median() <= 4
 
-        encodings = get_one_hot_encodings(sequences, dpm_patterns, constraints=[price_ct],
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns, constraints=[price_ct],
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([[1, 0, 0], [0, 1, 1], [1, 0, 1],
@@ -535,7 +303,8 @@ class TestDPMUtils(unittest.TestCase):
         price = Attribute(values=values_pos + values_neg)
         price_ct = 3 <= price.median() <= 4
 
-        encodings = get_one_hot_encodings(sequences, dpm_patterns, constraints=[price_ct],
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns, constraints=[price_ct],
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([[1], [0], [1],
@@ -584,7 +353,8 @@ class TestDPMUtils(unittest.TestCase):
         price = Attribute(values=values_pos + values_neg)
         price_ct = 3 <= price.median() <= 4
 
-        encodings = get_one_hot_encodings(sequences, dpm_patterns, constraints=[price_ct],
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns, constraints=[price_ct],
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([[0], [1], [1],
@@ -633,7 +403,8 @@ class TestDPMUtils(unittest.TestCase):
         price = Attribute(values=values_pos + values_neg)
         price_ct = 3 <= price.median() <= 4
 
-        encodings = get_one_hot_encodings(sequences, dpm_patterns, constraints=[price_ct],
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns, constraints=[price_ct],
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([[0], [1], [0],
@@ -678,7 +449,8 @@ class TestDPMUtils(unittest.TestCase):
                                                             min_frequency_pos=2, min_frequency_neg=2)
         dpm_patterns = aggregation_to_patterns[DichotomicAggregation.union]
 
-        encodings = get_one_hot_encodings(sequences, dpm_patterns,
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns,
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([[1, 1, 0], [0, 1, 1], [1, 0, 1],
@@ -744,7 +516,8 @@ class TestDPMUtils(unittest.TestCase):
 
         # Encodings of all sequences
         sequences = sequences_pos + sequences_neg
-        encodings = get_one_hot_encodings(sequences, dpm_patterns,
+        pat2feat = Pat2Feat()
+        encodings = pat2feat.get_features(sequences, dpm_patterns,
                                           drop_pattern_frequency=False)
 
         self.assertListEqual([['A', 'A'], ['A', 'A', 'A'], ['A', 'A', 'A', 'D'], ['A', 'A', 'B'],
